requirements.txt
dj-rest-auth==4.0.1
Django==4.2.7
django-allauth==0.50.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth[with_social]==4.0.1
gunicorn>20.0.0,<21.0
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
stripe==11.5.0